requirements.txt
# Update requirements.txt

requests==2.31.0
fastapi==0.104.1
uvicorn==0.24.0
kafka-python==2.0.2
psycopg2-binary==2.9.9
scikit-learn==1.3.1
numpy==1.24.3
pandas==2.1.2
python-dotenv==1.0.0
aiofiles==23.2.1
websockets==12.0
python-multipart==0.0.6
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.12.1
pyyaml==6.0.1
